ETQ Instructeur : je veux personnaliser l'affichage des badges de notification

+++ app/models/dossier_notification.rb
@@ -3,8 +3,15 @@
 class DossierNotification < ApplicationRecord
   DELAY_DOSSIER_DEPOSE = 7.days
 
-  belongs_to :instructeur
-  belongs_to :dossier
+  REFRESH_CONDITIONS_BY_TYPE = {
+    dossier_depose:          -> (dossier, _) { dossier.en_construction? && dossier.follows.empty? },
+    dossier_modifie:         -> (dossier, _) { dossier.last_champ_updated_at.present? && dossier.last_champ_updated_at > dossier.depose_at },
+    message:                 -> (dossier, instructeur_id) { dossier.commentaires.to_notify(instructeur_id).present? },
+    annotation_instructeur:  -> (dossier, _) { dossier.last_champ_private_updated_at.present? },
+    avis_externe:            -> (dossier, _) { dossier.avis.with_answer.present? },
+    attente_correction:      -> (dossier, _) { dossier.pending_correction? },
+    attente_avis:            -> (dossier, _) { dossier.avis.without_answer.present? }
+  }.freeze
 
   enum :notification_type, {
     dossier_depose: 'dossier_depose',
@@ -16,6 +23,9 @@ class DossierNotification < ApplicationRecord
     attente_avis: 'attente_avis'
   }
 
+  belongs_to :instructeur
+  belongs_to :dossier
+
   scope :to_display, -> { where(display_at: ..Time.current) }
 
   scope :order_by_importance, -> {
@@ -24,35 +34,46 @@ class DossierNotification < ApplicationRecord
 
   scope :type_news, -> { where(notification_type: [:dossier_modifie, :message, :annotation_instructeur, :avis_externe]) }
 
-  def self.create_notification(dossier, notification_type, instructeur: nil, except_instructeur: nil)
-    case notification_type
-    when :dossier_depose
-      if !dossier.procedure.declarative? && !dossier.procedure.sva_svr_enabled?
-        instructeur_ids = Array(instructeur&.id.presence || dossier.groupe_instructeur.instructeur_ids)
-        display_at = dossier.depose_at + DELAY_DOSSIER_DEPOSE
+  def self.create_notification(dossier, notification_type, except_instructeur: nil)
+    instructeur_ids = instructeur_to_notify_ids(dossier, notification_type, except_instructeur)
 
-        instructeur_ids.each do |instructeur_id|
-          find_or_create_notification(dossier, notification_type, instructeur_id, display_at:)
-        end
-      end
-
-    when :dossier_modifie, :message, :attente_correction, :attente_avis, :annotation_instructeur, :avis_externe
-      instructeur_ids = Array(instructeur&.id.presence || dossier.followers_instructeur_ids)
-      instructeur_ids -= [except_instructeur.id] if except_instructeur.present?
-
-      instructeur_ids.each do |instructeur_id|
-        find_or_create_notification(dossier, notification_type, instructeur_id)
-      end
+    instructeur_ids.each do |instructeur_id|
+      find_or_create_notification(dossier, notification_type, instructeur_id)
     end
   end
 
-  def self.refresh_notifications_instructeur_for_dossier(instructeur, dossier)
-    create_notification(dossier, :dossier_modifie, instructeur:) if dossier.last_champ_updated_at.present? && dossier.last_champ_updated_at > dossier.depose_at
-    create_notification(dossier, :message, instructeur:) if dossier.commentaires.to_notify(instructeur).present?
-    create_notification(dossier, :annotation_instructeur, instructeur:) if dossier.last_champ_private_updated_at.present?
-    create_notification(dossier, :avis_externe, instructeur:) if dossier.avis.with_answer.present?
-    create_notification(dossier, :attente_correction, instructeur:) if dossier.pending_correction?
-    create_notification(dossier, :attente_avis, instructeur:) if dossier.avis.without_answer.present?
+  def self.refresh_notifications_instructeur_for_dossier_by_choice(instructeur, dossier, choice)
+    instructeur_preferences = instructeur_preferences(instructeur, dossier.procedure)
+
+    notification_types_to_refresh = notification_types.keys.map(&:to_sym).filter do |notification_type|
+      instructeur_preferences[notification_type] == choice
+    end
+
+    return if notification_types_to_refresh.empty?
+
+    notification_types_to_refresh.each do |notification_type|
+      find_or_create_notification(dossier, notification_type, instructeur.id) if REFRESH_CONDITIONS_BY_TYPE[notification_type].call(dossier, instructeur.id)
+    end
+  end
+
+  def self.refresh_notifications_instructeur_for_followed_dossier(instructeur, dossier)
+    destroy_notifications_by_dossier_and_type(dossier, :dossier_depose)
+    refresh_notifications_instructeur_for_dossier_by_choice(instructeur, dossier, 'followed')
+  end
+
+  def self.refresh_notifications_instructeur_for_dossiers(all_dossiers, followed_dossiers, non_followed_dossiers, instructeur_id, notification_type, old_preference, new_preference)
+    case [old_preference, new_preference]
+    when ['all', 'none'], ['followed', 'none']
+      destroy_notifications_by_dossiers_and_type_and_instructeur(all_dossiers, notification_type, instructeur_id)
+    when ['all', 'followed']
+      destroy_notifications_by_dossiers_and_type_and_instructeur(non_followed_dossiers, notification_type, instructeur_id)
+    when ['none', 'all']
+      refresh_notifications_instructeur_for_dossiers_and_type(all_dossiers, notification_type, instructeur_id)
+    when ['followed', 'all']
+      refresh_notifications_instructeur_for_dossiers_and_type(non_followed_dossiers, notification_type, instructeur_id)
+    when ['none', 'followed']
+      refresh_notifications_instructeur_for_dossiers_and_type(followed_dossiers, notification_type, instructeur_id)
+    end
   end
 
   def self.destroy_notifications_instructeur_of_groupe_instructeur(groupe_instructeur, instructeur)
@@ -65,6 +86,20 @@ class DossierNotification < ApplicationRecord
   def self.destroy_notifications_instructeur_of_dossier(instructeur, dossier)
     DossierNotification
       .where(instructeur:, dossier:)
+      .destroy_all
+  end
+
+  def self.destroy_notifications_instructeur_of_unfollowed_dossier(instructeur, dossier)
+    instructeur_preferences = instructeur_preferences(instructeur, dossier.procedure)
+
+    notification_types_to_destroy = notification_types.keys.map(&:to_sym).reject do |notification_type|
+      instructeur_preferences[notification_type] == "all"
+    end
+
+    return if notification_types_to_destroy.empty?
+
+    DossierNotification
+      .where(instructeur:, dossier:, notification_type: notification_types_to_destroy)
       .destroy_all
   end
 
@@ -90,6 +125,7 @@ class DossierNotification < ApplicationRecord
       .distinct
 
     dossiers_with_news_notification_by_statut = {
+      a_suivre: dossiers_with_news_notification.by_statut('a-suivre'),
       suivis: dossiers_with_news_notification.by_statut('suivis', instructeur:),
       traites: dossiers_with_news_notification.by_statut('traites')
     }
@@ -108,6 +144,7 @@ class DossierNotification < ApplicationRecord
       .distinct
 
     {
+      a_suivre: dossiers_with_news_notification.by_statut('a-suivre').exists?,
       suivis: dossiers_with_news_notification.by_statut('suivis', instructeur:).exists?,
       traites: dossiers_with_news_notification.by_statut('traites').exists?
     }
@@ -211,16 +248,64 @@ class DossierNotification < ApplicationRecord
       .distinct
       .count
   end
-end
 
-private
+  private
 
-def find_or_create_notification(dossier, notification_type, instructeur_id, display_at: Time.current)
-  DossierNotification.find_or_create_by!(
-    dossier:,
-    notification_type:,
-    instructeur_id:
-  ) do |notification|
-    notification.display_at = display_at
+  def self.find_or_create_notification(dossier, notification_type, instructeur_id)
+    display_at = notification_type == :dossier_depose ? (dossier.depose_at + DossierNotification::DELAY_DOSSIER_DEPOSE) : Time.zone.now
+
+    DossierNotification.find_or_create_by!(
+      dossier:,
+      notification_type:,
+      instructeur_id:
+    ) do |notification|
+      notification.display_at = display_at
+    end
+  end
+
+  def self.instructeur_preferences(instructeur, procedure)
+    if (instructeur_procedure = InstructeursProcedure.find_by(instructeur:, procedure:))
+      instructeur_procedure.notification_preferences
+    else
+      InstructeursProcedure::DEFAULT_NOTIFICATIONS_PREFERENCES
+    end
+  end
+
+  def self.instructeur_to_notify_ids(dossier, notification_type, except_instructeur)
+    instructeur_ids = dossier.groupe_instructeur.instructeur_ids
+    instructeur_ids -= [except_instructeur.id] if except_instructeur.present?
+
+    followers_instructeur_ids = dossier.followers_instructeur_ids
+
+    preference_by_instructeur_id = InstructeursProcedure
+      .where(instructeur_id: instructeur_ids, procedure_id: dossier.procedure.id)
+      .pluck(:instructeur_id, "display_#{notification_type}_notifications")
+      .to_h
+
+    instructeur_ids.filter do |instructeur_id|
+      preference = preference_by_instructeur_id[instructeur_id] || InstructeursProcedure::DEFAULT_NOTIFICATIONS_PREFERENCES[notification_type]
+
+      if followers_instructeur_ids.include?(instructeur_id)
+        ['followed', 'all'].include?(preference)
+      else
+        preference == "all"
+      end
+    end
+  end
+
+  def self.destroy_notifications_by_dossiers_and_type_and_instructeur(dossiers, notification_type, instructeur_id)
+    DossierNotification
+      .where(dossier_id: dossiers, notification_type:, instructeur_id:)
+      .destroy_all
+  end
+
+  def self.refresh_notifications_instructeur_for_dossiers_and_type(dossiers, notification_type, instructeur_id)
+    missing_notifications = dossiers.filter_map do |dossier|
+      if DossierNotification::REFRESH_CONDITIONS_BY_TYPE[notification_type].call(dossier, instructeur_id)
+        display_at = notification_type == :dossier_depose ? (dossier.depose_at + DossierNotification::DELAY_DOSSIER_DEPOSE) : Time.zone.now
+        { dossier_id: dossier.id, instructeur_id:, notification_type:, display_at: }
+      end
+    end
+    DossierNotification.insert_all(missing_notifications) if missing_notifications.size.positive?
   end
 end
